fix: use typescript watcher for directories to prevent TS6307
Webpack watcher notifies only about changes of files that are in the
compilation dependencies graph. This means that it will not notice if we
create a new file that is not missing. This is a difference between
webpack and typescript watch mechanism that can lead to errors like
TS6307. This fix uses typescript watcher for directories.